Server was moved from LAN to the Internet. Also, design was improved
Now server works in the Internet. It still works in test mode, and machine which it works on is not turned on 24/7. Design of the app was upgraded to tell the user when the server is unavailable. Also, the app shows when the data is loading. Redundant MainActivity was removed. Now the ChartActivity is the launcher activity.