All application architecture was changed
In order to replace creating new threads in WSCLientEndpoint with coroutines all project architecture was changed to use ViewModel's coroutine scope (and, resppectively the ViewModel itself). Migration to ViewModel architecture also provoked migration to Navigation architecture. All of the activities were replaced by the fragments. Now it is a single-activity application with Navigation architecture which uses ViewModel and LiveData to share data between different parts of the app.